+++ plugins/zelda_mc/rules.py
@@ -197,9 +197,11 @@ class GameRules(BaseGameRules):
                             logical_hard_limit: Optional[int] = None,
                             allowed_problems: Optional[Set[str]] = None,
                             block_idx: Optional[int] = None,
-                            string_idx: Optional[int] = None) -> Tuple[str, bool]:
+                            string_idx: Optional[int] = None,
+                            page_local: bool = False,
+                            disable_pagination: bool = False) -> Tuple[str, bool]:
         return self.text_fixer.autofix_data_string(
-            data_string, editor_font_map, editor_line_width_threshold, logical_hard_limit, allowed_problems, block_idx, string_idx
+            data_string, editor_font_map, editor_line_width_threshold, logical_hard_limit, allowed_problems, block_idx, string_idx, page_local, disable_pagination
         )
     
     def process_pasted_segment(self,