+++ plugins/zelda_ww/rules.py
@@ -88,7 +88,7 @@ class GameRules(BaseGameRules):
             logical_hard_limit=logical_hard_limit
         )
 
-    def autofix_data_string(self, data_string: str, editor_font_map: dict, editor_line_width_threshold: int, logical_hard_limit: Optional[int] = None, allowed_problems: Optional[Set[str]] = None, block_idx: Optional[int] = None, string_idx: Optional[int] = None) -> Tuple[str, bool]:
+    def autofix_data_string(self, data_string: str, editor_font_map: dict, editor_line_width_threshold: int, logical_hard_limit: Optional[int] = None, allowed_problems: Optional[Set[str]] = None, block_idx: Optional[int] = None, string_idx: Optional[int] = None, page_local: bool = False, disable_pagination: bool = False) -> Tuple[str, bool]:
         return self.text_fixer.autofix_data_string(
             data_string=data_string,
             editor_font_map=editor_font_map,
@@ -96,7 +96,9 @@ class GameRules(BaseGameRules):
             logical_hard_limit=logical_hard_limit,
             allowed_problems=allowed_problems,
             block_idx=block_idx,
-            string_idx=string_idx
+            string_idx=string_idx,
+            page_local=page_local,
+            disable_pagination=disable_pagination
         )
 
     def process_pasted_segment(self, segment_to_insert: str, original_text_for_tags: str, editor_player_tag_const: str) -> Tuple[str, str, str]: